requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
autopep8==1.6.0
cffi==1.15.0
channels==3.0.4
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.20.0
pytz==2022.1
service-identity==21.1.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
uvicorn==0.17.6
zope.interface==5.4.0
